UI Complete

+++ src/main/java/model/Customer.java
@@ -1,19 +1,19 @@
 package model;
 
 public class Customer {
-    private int cellPhone;
+    private long cellPhone;
     private String name;
     private String address;
     private String dlicense;
 
-    public Customer(int cellPhone, String name, String address, String dlicense) {
+    public Customer(long cellPhone, String name, String address, String dlicense) {
         this.cellPhone = cellPhone;
         this.name = name;
         this.address = address;
         this.dlicense = dlicense;
     }
 
-    public int getCellPhone() {
+    public long getCellPhone() {
         return cellPhone;
     }
